Tied django-livesettings commit
I haven't fully investigated this yet but a couple of people have
reported a RECURSIVE_RELATIONSHIP_CONSTANT error in actionitems.

This seems to be due in part to an incompatibility with the latest
version of django-livesettings, which is generating an error about
config_register not being imported in publicweb/config.py. At some
point we're going to need to change this to:

from livesettings.functions import ...

The quickest fix is to tie our version of livesettings to a working
commit. This is probably good practice anyway.